Get locations

+++ GroupAPIProject.Services/Location/LocationService.cs
@@ -33,8 +33,9 @@ namespace GroupAPIProject.Services.Location
             LocationEntity locationEntity = new LocationEntity
             {
 
-                LocationName = request.LocationName
-                RetailerId = _retailerId,
+                LocationName = request.LocationName,
+                Capacity = request.Capacity,
+                RetailerId = _retailerId
 
             };
             _context.Locations.Add(locationEntity);
@@ -42,16 +43,33 @@ namespace GroupAPIProject.Services.Location
             return numberOfChanges == 1;
         }
 
-        public async Task<bool> RemoveLocationAsync(string LocationName)
+        public async Task<bool> RemoveLocationAsync(int LocationId)
         {
-            var locationEntity = await _context.Locations.FirstOrDefaultAsync(s => s.LocationName == LocationName);
+            var locationEntity = await _context.Locations.Where(entity => entity.RetailerId == _retailerId).FirstOrDefaultAsync(s => s.Id == LocationId);
 
             if (locationEntity == null)
             {
                 return false;
             }
+            if (locationEntity.ListOfInventoryItems.Count == 0)
+            {
+                return false;
+            }
             _context.Locations.Remove(locationEntity);
             return await _context.SaveChangesAsync() == 1;
+        }
+
+        public async Task<IEnumerable<LocationDetail>> GetLocationListAsync()
+        {
+            var LocationToDisplay = await _context.Locations.Where(entity => entity.RetailerId == _retailerId)
+                .Select(entity => new LocationDetail
+                {
+                    Id = entity.Id,
+                    LocationName = entity.LocationName,
+                    Capacity = entity.Capacity
+                }).ToListAsync();
+            
+            return LocationToDisplay;
         }
     }
 }
